Adding GIF, refactoring similar code out to comon class method

+++ sort/sort_JPEG.py
@@ -2,11 +2,8 @@
 import logging
 import os
 import shutil
-import pprint
 
 from .readExif import readExifData
-
-pp = pprint.PrettyPrinter(indent=4)
 
 class sort( readExifData ):
 
@@ -17,14 +14,7 @@ class sort( readExifData ):
       self.exif_data = exif_data
 
    def main(self):
-      create_date = None
-      if 'EXIF:DateTimeOriginal' in self.exif_data:
-         create_date = self.exif_data['EXIF:DateTimeOriginal']
-      elif 'File:FileModifyDate' in self.exif_data:
-         self.logger.info("No create date found, using file modify date")
-         create_date = self.exif_data['File:FileModifyDate']
-      else:
-         self.logger.error("Something is wrong no time stamp found")
+      create_date = self.get_create_date(self.exif_data)
       
       splitDate = create_date.split(" ")
       date = splitDate[0]
@@ -40,6 +30,8 @@ class sort( readExifData ):
       dest_file = ("/".join(date_parts))
       self.logger.debug("Sorted destination file: {}".format(dest_file))
 
-      os.makedirs(dest_dir, exist_ok=True)
-      shutil.copyfile(self.exif_data['SourceFile'], dest_file)
+      if self.args.dry_run:
+         os.makedirs(dest_dir, exist_ok=True)
+         shutil.copyfile(self.exif_data['SourceFile'], dest_file)
+      
       self.logger.info("Copied File: {} into {}".format(self.exif_data['SourceFile'], dest_dir))
